More to do...

+++ rules/games/piece.h
@@ -15,16 +15,16 @@ namespace game{
 	class Piece{
 	public:
 		Piece(const size_t id): _id(id){}
-		Piece(const size_t id, int x, int y, int type, int owner, int status):
+		Piece(const size_t id, size_t x, size_t y, int type, int owner, int status):
 			_id(id), _x(x), _y(y), _type(type), _owner(owner), _status(status) {}
 		
-		void init(int x, int y, int type, int owner, int status){
+		void init(size_t x, size_t y, int type, int owner, int status){
 			_x = x; _y = y;
 			_type = type;
 			_owner = owner;
 			_status = status;
 		}
-		void setPos(int x, int y) { _x = x; _y = y; }
+		void setPos(size_t x, size_t y) { _x = x; _y = y; }
 		void setOwner(int owner) { _owner = owner; }
 		void setStatus(int status) { _status = status; }
 
@@ -34,8 +34,8 @@ namespace game{
 		int _owner;
 		int _status;
 		
-		int _x;
-		int _y;
+		size_t _x;
+		size_t _y;
 
 	};
